scan: Allow pruned iDCT to be used with offset scans
FFTW will now be used only for the larger one-time iDCT of coefficients
behind the offset if the pruned iDCT would be faster for the rest of the
scan.

Also fixes a bug where FFTW was unintentionally being selected for all
scans that didn't use the --offset and --skip options, even with
--pruned-idct true.

Should provide a modest speedup for most scans with smaller intervals.