precision: handle defined but unset macros more broadly
Slightly reworks the change made in 193595b39af2fc7b9a2390ba09e89b8c5d89572c
so that `-DINTERMEDIATE_PRECISION=` and `-DINTERMEDIATE_PRECISION=0` are
both treated as if INTERMEDIATE_PRECISION was not provided, in which
case defaults are used. This is expanded to the COEFF_PRECISION macro
as well.