ffapi: add mechanism to specify supported input pixel formats
Adds a callback to ffapi_open_input which is given an AVPixFmtDescriptor
and should return a boolean value indicating whether the application can
process that pixel format.

ffapi will use this to check if the pixel format of a given input is
supported and if not to find the closest supported format to convert to.
However, if a specific pixel format is requested via color_props that
isn't supported according to this callback this is treated as an error.

One callbacks is defined by the library to select only 8 bit formats
compatible with ffapi_getpel/setpel and is used by all dspfun apps, but
rotate further limits this to only formats with no subsampling.

As a result, rotate no longer requires a pixel_format be explicitly
specified for subsampled input; this will be detected and converted
automatically.