spec: refactor shared option parsing
Makes calls to getopt the responsibility of the individual application
rather than the shared option parsing function.

This is a bit cleaner and lets us avoid allocating memory for the usage
text that was previously left to be freed by the OS on exit.

With this change spec can now properly exit with an error if an invalid
option is given.