scan: add option to measure parity with the input image
This prints the scan index at which the reconstructed image becomes
identical to the original, which may be before the end of the scan
particularly when there is less high-frequency content in the image.

Can be used as a measure of efficiency for the different scan methods,
with those that converge on the input earlier being more effective than
others.