motion: normalize coefficients to uniform range for processing
This was previously done for quantization, but other operations stand to
benefit from it, e.g. spectrograms will appear more uniform.

Neither is more or less correct per se, so this was initially omitted
for performance, but in practice there appears to be little to no
impact.

Note that this change potentially affects the results of the following
operations for the same input:
--coeff-limit, --threshold, --eval, --(i)spectrogram