ffapi: forward errors from ffapi_read_frame in ffapi_seek_frame
ffapi_seek_frame will now return any errors during frame reading.

Previously this returned the total number of frames seeked, which may
be less than `offset` in case there are fewer frames in the input than
requested.
Instead `offset` is now an in/out param which will be updated with the
total; this may be more convenient for applications doing calculations
with the offset after seeking as well.