ffapi: add specific error output to ffapi_open_*
Adds an optional out param `averror` to ffapi_open_input and
ffapi_open_output which if provided will be set to either 0 on success
or an AVERROR code on failure. This can be used to obtain a descriptive
error message with e.g. av_err2str.

This forwards all errors from the FFmpeg library calls as well as any
other condition that would cause these functions to return a NULL
FFContext.